change topbarheight into int

+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/menu/customize/MenuCustomizeShortcutsScreen.java
@@ -67,7 +67,7 @@ public class MenuCustomizeShortcutsScreen extends MenuBackground {
 		float boxWidth = (float) width * 0.67F;
 		float middleHeight = (float) height * 0.6F;
 
-		buttonPosY = topBarHeight + 5;
+		buttonPosY = (int) topBarHeight + 5;
 		
 		box = new MenuBox((int) boxPosX, (int) topBarHeight, (int) boxWidth, (int) middleHeight, new Color(4, 4, 68));
 		buttonsX = box.getX() + 10;
